+++ spreadsheet-poi/src/main/java/ec/util/spreadsheet/poi/PoiBookWriter.java
@@ -1,31 +1,28 @@
 /*
  * Copyright 2013 National Bank of Belgium
  *
- * Licensed under the EUPL, Version 1.1 or – as soon they will be approved 
+ * Licensed under the EUPL, Version 1.1 or – as soon they will be approved
  * by the European Commission - subsequent versions of the EUPL (the "Licence");
  * You may not use this work except in compliance with the Licence.
  * You may obtain a copy of the Licence at:
  *
  * http://ec.europa.eu/idabc/eupl
  *
- * Unless required by applicable law or agreed to in writing, software 
+ * Unless required by applicable law or agreed to in writing, software
  * distributed under the Licence is distributed on an "AS IS" basis,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the Licence for the specific language governing permissions and 
+ * See the Licence for the specific language governing permissions and
  * limitations under the Licence.
  */
 package ec.util.spreadsheet.poi;
 
-import java.io.IOException;
-import org.apache.poi.ss.usermodel.Cell;
-import org.apache.poi.ss.usermodel.CellStyle;
-import org.apache.poi.ss.usermodel.Row;
-import org.apache.poi.ss.usermodel.Sheet;
-import org.apache.poi.ss.usermodel.Workbook;
+import org.apache.poi.ss.usermodel.*;
 import org.apache.poi.xssf.streaming.SXSSFSheet;
 
+import java.io.IOException;
+import java.util.Date;
+
 /**
- *
  * @author Philippe Charles
  */
 final class PoiBookWriter {
@@ -58,9 +55,9 @@ final class PoiBookWriter {
         for (int i = 0; i < source.getRowCount(); i++) {
             Row row = target.createRow(i);
             for (int j = 0; j < source.getColumnCount(); j++) {
-                ec.util.spreadsheet.Cell cell = source.getCell(i, j);
-                if (cell != null) {
-                    copy(cell, row.createCell(j));
+                Object cellValue = source.getCellValue(i, j);
+                if (cellValue != null) {
+                    copyCellValue(cellValue, row.createCell(j));
                 }
             }
         }
@@ -72,14 +69,14 @@ final class PoiBookWriter {
         }
     }
 
-    private void copy(ec.util.spreadsheet.Cell source, Cell target) {
-        if (source.isDate()) {
-            target.setCellValue(source.getDate());
+    private void copyCellValue(Object cellValue, Cell target) {
+        if (cellValue instanceof Date) {
+            target.setCellValue((Date) cellValue);
             target.setCellStyle(dateStyle);
-        } else if (source.isNumber()) {
-            target.setCellValue(source.getDouble());
-        } else if (source.isString()) {
-            target.setCellValue(source.getString());
+        } else if (cellValue instanceof Number) {
+            target.setCellValue(((Number) cellValue).doubleValue());
+        } else if (cellValue instanceof String) {
+            target.setCellValue((String) cellValue);
         }
     }
 }